UX-437 Convert utility stories to libby
