--- libby/utility/usePrefersReducedMotion.stories.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import { Box, usePrefersReducedMotion } from '@sparkpost/matchbox';
-
-export default {
-  title: 'Utility|usePrefersReducedMotion',
-};
-
-export const ExampleUsage = () => {
-  const prefersReducedMotion = usePrefersReducedMotion();
-
-  return (
-    <>
-      <Box>(prefers-reduced-motion: {prefersReducedMotion})</Box>
-    </>
-  );
-};
